Fix ES community_id mapping + update docs (memory + indexing)

+++ apps/api/services/elasticsearch_service.py
@@ -102,7 +102,7 @@ class ElasticsearchService:
                     "venue_name": {"type": "text"},
                     "pagerank": {"type": "float"},
                     "citation_count": {"type": "integer"},
-                    "community_id": {"type": "integer"}
+                    "community_id": {"type": "long"}
                 }
             }
         }
